requirements.txt
moderngl
moderngl-window
numba
numpy
PyGLM
pygltflib
scipy
